refactor: move drawing and screenshot functions, move HandleCommand type
