+++ src/commandHandlers.ts/drawRectangle.ts
