+++ src/commandHandlers.ts/drawSquaare.ts
