+++ src/utils.ts/isCommandValid.ts
@@ -4,16 +4,7 @@ import {
   DRAW_SQUARE,
   MOUSE_DOWN, MOUSE_LEFT, MOUSE_POSITION, MOUSE_RIGHT, MOUSE_UP, PRINT_SCREEN,
 } from '../const';
-
-type HandlerCommand = typeof MOUSE_UP
-  | typeof MOUSE_DOWN
-  | typeof MOUSE_LEFT
-  | typeof MOUSE_RIGHT
-  | typeof MOUSE_POSITION
-  | typeof DRAW_CIRCLE
-  | typeof DRAW_RECTANGLE
-  | typeof DRAW_SQUARE
-  | typeof PRINT_SCREEN;
+import { HandlerCommand } from '../types/utils';
 
 function isCommandValid(command: string | HandlerCommand): command is HandlerCommand {
   return (command === MOUSE_UP
